Shift test cases around
