+++ boosters/node-beginner/meeting-4/src/api/routes/repos/index.js
@@ -3,5 +3,6 @@ const router = express.Router();
 const { repoController } = require("../../controllers");
 
 router.route("/").get(repoController.getAllRepos);
+router.get("/:repoId", repoController.getRepo);
 
 module.exports = router;
