+++ boosters/node-beginner/meeting-4/tests/00_exercise.test.js
@@ -1,35 +1,31 @@
 const request = require("supertest");
 const app = require("../src/server");
-const setup = require("./config/setup");
 const teardown = require("./config/teardown");
 const uuid = require("uuid/v4");
+const setup = require("./config/setup");
 
-const users = [
-  { firstName: "Andrew", lastName: "Mayer", id: uuid() },
-  { firstName: "Liang", lastName: "Chen", id: uuid() },
+const user = [
+  {
+    firstName: "Andrew",
+    lastName: "Mayer",
+    id: uuid(),
+  },
 ];
 
-const repos = users.reduce(
-  (acc, curr) => [
-    ...acc,
-    { name: "repo1", userId: curr.id, isPrivate: false, id: uuid() },
-    { name: "repo2", userId: curr.id, isPrivate: true, id: uuid() },
-  ],
-  []
-);
+const repo = {
+  name: "repo1",
+  userId: user.id,
+  isPrivate: false,
+};
 
-describe("User Repos", () => {
-  beforeEach(async () => {
-    await setup("Users", users)();
-    return setup("Repos", repos)();
-  });
-  afterEach(teardown);
-  it("Should return all repos of a given userId", async () => {
-    res = await Promise.all(users.map(user => request(app).get(`/users/${user.id}/repos`)));
+describe("Create Repo", () => {
+  beforeEach(setup("Users", user));
+  afterAll(teardown);
 
-    res.map(res => {
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.length).toEqual(2);
-    });
+  xit("Should create a repo", async () => {
+    const res = await request(app)
+      .post("/repos/")
+      .send(repo);
+    expect(res.statusCode).toEqual(201);
   });
 });
